Initialize arch-dir and move network arch in network dir
