--- client/client.cmake
@@ -1,1 +1,0 @@
-set(client_SOURCE_FILES )
