+++ client/main_client.cpp
@@ -1,0 +1,12 @@
+//
+//-----------------------------------------------------------------------------
+//Filename:    main_client.cpp
+//-----------------------------------------------------------------------------
+//
+// Created by bonett_w on 1/9/18.
+//-----------------------------------------------------------------------------
+//
+
+int main(int argc, char *argv[]) {
+    return (0);
+}
