--- core/core.cmake
@@ -1,1 +1,0 @@
-set(core_SOURCE_FILES )
