--- engine/game/game.cmake
@@ -1,1 +1,0 @@
-set(game_SOURCE_FILES )
