--- engine/graphics/graphics.cmake
@@ -1,1 +1,0 @@
-set(graphics_SOURCE_FILES )
