--- engine/network/network.cmake
@@ -1,1 +1,0 @@
-set(network_SOURCE_FILES )
