--- engine/physics/physics.cmake
@@ -1,2 +1,0 @@
-
-set(physics_SOURCE_FILES )
