+++ server/main_server.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "Socket.hpp"
+#include "../engine/network/Socket.hpp"
 
 int main() {
     RType::Socket *socket = new RType::Socket();
