--- server/server.cmake
@@ -1,1 +1,0 @@
-set(server_SOURCE_FILES )
